4.32. Desafio implementando serviços REST de cidades e estados

+++ src/main/java/com/algaworks/algafood/repository/CidadeRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,8 +35,11 @@ public class CidadeRepositoryImpl implements CidadeRepository {
 	
 	@Override
 	@Transactional
-	public void remover(Cidade cidade) {
-		cidade = buscar(cidade.getId());
+	public void remover(Long id) {
+		Cidade cidade = buscar(id);
+		if (cidade == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
 		entityManager.remove(cidade);
 	}
