+++ src/main/java/com/algaworks/algafood/repository/EstadoRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,8 +35,11 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 	
 	@Override
 	@Transactional
-	public void remover(Estado estado) {
-		estado = buscar(estado.getId());
+	public void remover(Long id) {
+		Estado estado = buscar(id);
+		if (estado == null) {
+			throw new EmptyResultDataAccessException(1);
+		}		
 		entityManager.remove(estado);
 	}
